+++ src/main/java/com/github/ssullivan/db/redis/search/FindBySearchRequest.java
@@ -1,208 +1,267 @@
 package com.github.ssullivan.db.redis.search;
 
 import static com.github.ssullivan.db.redis.RedisConstants.DEFAULT_EXPIRE_SECONDS;
+import static com.github.ssullivan.db.redis.RedisConstants.indexByGeoKey;
 
 import com.github.ssullivan.db.IFacilityDao;
 import com.github.ssullivan.db.IFeedDao;
-import com.github.ssullivan.db.IFindBySearchRequest;
 import com.github.ssullivan.db.redis.IAsyncRedisConnectionPool;
 import com.github.ssullivan.db.redis.IRedisConnectionPool;
 import com.github.ssullivan.db.redis.ToFacilityWithRadiusConverter;
 import com.github.ssullivan.model.AvailableServices;
 import com.github.ssullivan.model.Facility;
+import com.github.ssullivan.model.FacilityWithRadius;
+import com.github.ssullivan.model.GeoPoint;
+import com.github.ssullivan.model.GeoRadiusCondition;
+import com.github.ssullivan.model.GeoUnit;
 import com.github.ssullivan.model.Page;
 import com.github.ssullivan.model.SearchRequest;
 import com.github.ssullivan.model.SearchResults;
 import com.github.ssullivan.model.ServicesCondition;
+import com.github.ssullivan.model.SetOperation;
+import com.github.ssullivan.model.collections.Tuple2;
 import com.github.ssullivan.utils.ShortUuid;
-import com.spotify.futures.CompletableFutures;
+import io.lettuce.core.GeoRadiusStoreArgs;
 import io.lettuce.core.ZStoreArgs;
-import io.lettuce.core.api.async.RedisAsyncCommands;
+import io.lettuce.core.api.StatefulRedisConnection;
+import io.lettuce.core.api.sync.RedisCommands;
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import java.util.function.BiFunction;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 import javax.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class FindBySearchRequest extends AbstractFindFacility implements IFindBySearchRequest {
+public class FindBySearchRequest extends AbstractFindFacility {
   private static final Logger LOGGER = LoggerFactory.getLogger(FindBySearchRequest.class);
 
   @Inject
   public FindBySearchRequest(IRedisConnectionPool redis,
       IAsyncRedisConnectionPool asyncPool,
-      IFacilityDao facilityDao,
-      IFeedDao feedDao) {
+      IFacilityDao facilityDao, IFeedDao feedDao) {
     super(redis, asyncPool, facilityDao, feedDao);
   }
-
 
   public CompletionStage<SearchResults<Facility>> find(SearchRequest searchRequest,
       Page page) throws Exception {
 
-    if (null == searchRequest) {
-      LOGGER.warn("Invalid Search Request of null. No Results!");
-      return CompletableFuture.completedFuture(SearchResults.searchResults(0));
-    }
+    Long totalResults = 0L;
+    List<String> facilityIdentifiers = new ArrayList<>();
 
+    try (StatefulRedisConnection<String, String> connection = syncPool.borrowConnection()) {
+      RedisCommands<String, String> sync = connection.sync();
 
-    return asyncPool.borrowConnection()
-        .thenCompose(connection -> {
-          try {
-            return find(connection.async(), searchRequest, page)
-                .whenComplete((s, throwable) -> asyncPool.relase(connection));
-          } catch (IOException e) {
-            e.printStackTrace();
-          }
+      final Optional<String> feedOption = this.feedDao.searchFeedId();
 
-          return CompletableFuture.completedFuture(SearchResults.searchResults(0, new ArrayList<>()));
-        });
-  }
+      if (!feedOption.isPresent()) {
+        LOGGER.error("Error connecting to the backend database! Unable to find current feed");
+        return CompletableFuture.completedFuture(SearchResults.empty());
+      }
+      final String searchKey = "s:" + ShortUuid.randomShortUuid() + ":";
 
-  private CompletionStage<SearchResults<Facility>> find(RedisAsyncCommands<String, String> async, SearchRequest searchRequest, Page page)
-      throws IOException {
+      final Tuple2<Long, String> serviceCodeResults =
+          findByServiceCodeConditions(sync, searchKey, searchRequest.getFinalSetOperation(), searchRequest.getConditions());
 
+      LOGGER.debug("Found {} matching locations for the service code conditions", serviceCodeResults.get_1());
 
-    final Optional<String> feedOption = this.feedDao.currentFeedId();
+      totalResults = serviceCodeResults.get_1();
+      if (searchRequest.getGeoRadiusCondition() != null) {
+        final Tuple2<Long, String> geoResult = findByGeoPoint(sync, feedOption.get(),
+            searchKey, searchRequest.getGeoRadiusCondition());
 
-    if (!feedOption.isPresent()) {
-      LOGGER.error("Error connecting to the backend database! Unable to find current feed");
-      return CompletableFuture.completedFuture(SearchResults.empty());
-    }
-    final String searchKey = "s:" + ShortUuid.randomShortUuid() + ":";
-    final String resultKey = searchKey + ":0";
-    final String geoKey = searchKey + ":geo";
+        LOGGER.debug("");
 
-    async.setAutoFlushCommands(false);
-
-
-    // Perform the following operations in Batch
-    boolean hasResultKey = false;
-    final List<ServicesCondition> servicesConditions = searchRequest.getConditions();
-
-    if (!servicesConditions.isEmpty()) {
-      int counter = 0;
-
-      final String[] keys = new String[servicesConditions.size()];
-      for (final ServicesCondition servicesCondition : servicesConditions) {
-        final String key = searchKey + "s:" + counter;
-        if (!createServiceSearchSet(async, key, feedOption.orElse(""), servicesCondition)) {
-          LOGGER.warn("Failed to create search set: {}", key);
+        if (serviceCodeResults.get_1() > 0) {
+          totalResults = sync.zinterstore(serviceCodeResults.get_2(), serviceCodeResults.get_2(),
+              geoResult.get_2());
         }
-        keys[counter] = key;
-        counter = counter + 1;
+        else {
+          totalResults = sync.zunionstore(serviceCodeResults.get_2(), serviceCodeResults.get_2(),
+              geoResult.get_2());
+        }
       }
 
-      switch (searchRequest.getFinalSetOperation()) {
+      if (totalResults == null) {
+        totalResults = 0L;
+      }
+
+      final List<String> results = sync
+          .zrange(serviceCodeResults.get_2(), page.offset(), page.offset() + page.size());
+      facilityIdentifiers.addAll(results);
+    }
+
+
+    final Long totalFound = totalResults;
+
+    final Function<List<Facility>, List<Facility>> toFacilityWithRadius =
+        applyToFacilityWithRadius(searchRequest);
+
+
+    return this.facilityDao.fetchBatchAsync(facilityIdentifiers)
+        .thenApply(this::applyAvailableServicesAll)
+        .thenApply(toFacilityWithRadius)
+        .thenApply(it -> SearchResults.searchResults(totalFound, it));
+  }
+
+  private Tuple2<Long, String> findByGeoPoint(final RedisCommands<String, String> sync,
+      final String feedKey,
+      final String searchKey,
+      final GeoRadiusCondition geoRadiusCondition) {
+
+    Objects.requireNonNull(sync, "Redis connection must not be null");
+    Objects.requireNonNull(searchKey, "Search key must not be null");
+    Objects.requireNonNull(geoRadiusCondition, "Geo Condition  must not be null");
+
+
+    Long totalResults = sync
+        .georadius(indexByGeoKey(feedKey), geoRadiusCondition.getGeoPoint()
+                .lon(), geoRadiusCondition.getGeoPoint().lat(),
+            geoRadiusCondition.getRadius(), geoRadiusCondition.getGeoUnit().unit(),
+            GeoRadiusStoreArgs.Builder
+                .withStoreDist(searchKey + "geo"));
+
+    if (totalResults == null) {
+      totalResults = 0L;
+    }
+    LOGGER.debug("Found {} locations within {} {} of {}", totalResults, geoRadiusCondition.getRadius(),
+        geoRadiusCondition.getGeoUnit(), geoRadiusCondition.getGeoPoint());
+
+    sync.expire(searchKey + "geo", DEFAULT_EXPIRE_SECONDS);
+    return new Tuple2<>(totalResults, searchKey + "geo");
+  }
+
+  /**
+   * Find locations by the provided {@link ServicesCondition}'s.
+   *
+   * @param sync connection to redis
+   * @param searchKey the search key prefix
+   * @param setOperation an final set operation to perfrom across all the result sets
+   * @param conditions the user provided search conditions to filter by
+   * @return an instance of {@link Tuple2} wher the first item is the number of results, and the
+   *         second result is the key where the final results are stored
+   */
+  private Tuple2<Long, String> findByServiceCodeConditions(final RedisCommands<String, String> sync,
+      final String searchKey,
+      final SetOperation setOperation,
+      final List<ServicesCondition> conditions) throws IOException {
+
+    Objects.requireNonNull(sync, "Redis connection must not be null");
+    Objects.requireNonNull(searchKey, "Search key must not be null");
+    Objects.requireNonNull(setOperation, "Set operation must not be null");
+    Objects.requireNonNull(conditions, "ServicesConditions must not be null");
+
+
+    final String[] conditionKeys = new String[conditions.size()];
+    int i = 0;
+    for (ServicesCondition servicesCondition : conditions) {
+      conditionKeys[i] = searchKey + i;
+      long totalFound = findByServiceConditions(sync, conditionKeys[i], servicesCondition);
+      i++;
+    }
+
+
+    Long totalFound = 0L;
+    if (!conditions.isEmpty()) {
+      switch (setOperation) {
         case INTERSECTION:
-          async.sinterstore(resultKey, keys);
+          totalFound = sync.zinterstore(searchKey + ":f", conditionKeys);
+          break;
         case UNION:
-          async.sunionstore(resultKey, keys);
+          totalFound = sync.zunionstore(searchKey + ":f", conditionKeys);
         default:
-          async.sinterstore(resultKey, keys);
       }
 
-      async.del(keys);
-      hasResultKey = true;
+      sync.del(conditionKeys);
     }
 
-    final boolean hasGeoSet = createGeoSearchSet(async,
-        geoKey,
-        feedOption.orElse(""),
-        searchRequest.getGeoRadiusCondition());
-
-    CompletionStage<Long> totalSearchResultsFuture = CompletableFuture.completedFuture(0L);
-    if (hasResultKey && hasGeoSet) {
-
-      async.zinterstore(searchKey, ZStoreArgs.Builder.weights(1.0), resultKey);
-
-      // #3 Find the intersection of the places that have our services we want and
-      //    are within a specific radius
-      totalSearchResultsFuture = async.zinterstore(searchKey, searchKey, geoKey);
-      async.expire(searchKey, DEFAULT_EXPIRE_SECONDS);
-      async.del(resultKey, geoKey);
+    if (totalFound == null) {
+      totalFound = 0L;
     }
-    else if (hasResultKey) {
-      totalSearchResultsFuture = async
-          .zinterstore(searchKey, ZStoreArgs.Builder.weights(1.0), resultKey);
-      async.expire(searchKey, DEFAULT_EXPIRE_SECONDS);
-      async.del(resultKey);
-    }
-    else if (hasGeoSet) {
-      totalSearchResultsFuture = async.zinterstore(searchKey, geoKey);
-      async.expire(searchKey, DEFAULT_EXPIRE_SECONDS);
-      async.del(geoKey);
-    }
-
-    async.flushCommands();
-    async.setAutoFlushCommands(true);
-
-
-    // #4 Fetch the results
-    final CompletionStage<List<String>> idFutures = async
-        .zrange(searchKey, page.offset(), page.offset() + page.size());
-    async.getStatefulConnection().setAutoFlushCommands(true);
-
-    final FetchFacilities fetchFacilities = new FetchFacilities(searchRequest, facilityDao, hasGeoSet);
-
-    return CompletableFutures.combineFutures(idFutures, totalSearchResultsFuture, fetchFacilities::apply)
-        .whenComplete((result, error) -> {
-          if (error != null) {
-            LOGGER.error("An error occurred while processing search", error);
-          }
-          // #6 Explicitly delete the keys
-          async.del(searchKey);
-          async.getStatefulConnection().setAutoFlushCommands(true);
-          async.getStatefulConnection().close();
-        });
+    return new Tuple2<>(totalFound, searchKey + ":f");
   }
 
-  private static class FetchFacilities implements BiFunction<Collection<String>, Long, CompletionStage<SearchResults<Facility>>> {
-    private IFacilityDao facilityDao;
-    private SearchRequest searchRequest;
-    private boolean hasGeoSet;
 
-    FetchFacilities(SearchRequest searchRequest, IFacilityDao facilityDao, boolean hasGeoSet) {
-      this.searchRequest = searchRequest;
-      this.hasGeoSet = hasGeoSet;
-      this.facilityDao = facilityDao;
+  private long findByServiceConditions(final RedisCommands<String, String> sync, final String searchKey, final ServicesCondition servicesCondition)
+      throws IOException {
+    /**
+     * Locations are are indexed by service code. There are multiple redis sets
+     * organized like servicecode -> [ facility 1, facility 2, etc.. ]
+     *
+     * Based on the user's query we are figuring which of these service code indices we need to
+     * query.
+     */
+    final String[] services = getServiceCodeIndices(feedDao.searchFeedId().orElse(""), servicesCondition.getServices());
+    Long totalFound = 0L;
+    switch (servicesCondition.getMatchOperator()) {
+      case MUST:
+        totalFound = sync.sinterstore(searchKey, services);
+        break;
+      case SHOULD:
+        totalFound = sync.sunionstore(searchKey, services);
+        break;
+      default:
+    }
+    sync.expire(searchKey, DEFAULT_EXPIRE_SECONDS);
+
+    if (totalFound == null) {
+      totalFound = 0L;
     }
 
+    LOGGER.debug("[{}] Found {} services matching {}", searchKey, totalFound, servicesCondition.getServices());
 
-    @Override
-    public CompletionStage<SearchResults<Facility>> apply(Collection<String> ids, Long totalResults) {
-      return facilityDao.fetchBatchAsync(ids)
-          .thenApply(facilities -> facilities
-              .stream()
-              .peek(facility -> {
-                final AvailableServices availableServices = facilityDao.getAvailableServices(facility);
-                facility.setAvailableServices(availableServices);
-              })
-              .map(facility -> {
-                if (hasGeoSet) {
-                  final double latitude = searchRequest.getGeoRadiusCondition().getGeoPoint()
-                      .lat();
-                  final double longitude = searchRequest.getGeoRadiusCondition().getGeoPoint()
-                      .lon();
-                  final String geoUnit = searchRequest.getGeoRadiusCondition().getGeoUnit()
-                      .getAbbrev();
+    // searchKey now contains a all of the ids for the locations we are looking for
+    final String[] mustNotServices = getServiceCodeIndices(servicesCondition.getMustNotServiceCodes());
+    Long totalDiff = 0L;
+    if (totalFound > 0 && mustNotServices.length > 0) {
+      sync.sunionstore(searchKey + ":!", mustNotServices);
+      totalDiff = sync.sdiffstore(searchKey, searchKey, searchKey + ":!");
+      sync.expire(searchKey + ":!", DEFAULT_EXPIRE_SECONDS);
+      sync.del(searchKey + ":!");
 
+      if (totalDiff == null) {
+        totalDiff = 0L;
+      }
 
-                  final ToFacilityWithRadiusConverter converter = new ToFacilityWithRadiusConverter(latitude, longitude, geoUnit);
-                  return converter.apply(facility);
-                }
-
-                return facility;
-              })
-              .collect(Collectors.toList()))
-          .thenApply(facilities -> SearchResults.searchResults(totalResults, facilities));
+      LOGGER.debug("[{}] Found {} services matching {} after diff", searchKey, totalDiff, servicesCondition.getServices());
+      return totalDiff;
     }
+
+    return totalFound;
   }
 
+
+  private List<Facility> applyAvailableServicesAll(final List<Facility> facilities) {
+    for (Facility facility : facilities) {
+      applyAvailableServices(facility);
+    }
+    return facilities;
+  }
+
+  private Facility applyAvailableServices(final Facility facility) {
+    final AvailableServices availableServices = facilityDao.getAvailableServices(facility);
+    facility.setAvailableServices(availableServices);
+    return facility;
+  }
+
+  /**
+   * Higher order function to return a function that will change Facility to FacilityWithRadius.
+   *
+   * @param searchRequest
+   * @return
+   */
+  private Function<List<Facility>, List<Facility>> applyToFacilityWithRadius(final SearchRequest searchRequest) {
+
+    final GeoPoint geoPoint = searchRequest.getGeoRadiusCondition() != null ? searchRequest.getGeoRadiusCondition().getGeoPoint() : null;
+    final GeoUnit geoUnit = searchRequest.getGeoRadiusCondition() != null ? searchRequest.getGeoRadiusCondition().getGeoUnit() : GeoUnit.MILE;
+
+    final ToFacilityWithRadiusConverter ToFacilityWithRadius = new ToFacilityWithRadiusConverter(geoPoint, geoUnit.getAbbrev());
+
+    return facilities1 -> facilities1.stream().map(ToFacilityWithRadius).collect(Collectors.toList());
+  }
 }
